Updated controller and test code. Updated generated code.
Added HTML for the controller.

Added another book to the mock data with a note about why we're using
mock data.

Added code for searching through book titles.

Added $watch for searching and sorting $scope variables with an
explanation on how the $watch function works.

Renamed $scope to scope in the tests.

Fixed how the existence of the controller is checked in the test.

Added $apply calls all over the tests. Added more tests.